Initial implementation of XML output
This is an incomplete implementation of XML output. We need to recursively
walk the tree of metrics, so the </whatever> appears in the right
place. So type Metrix is now hierarchical (each Metrix contains the Metrix
for nested nodes).  XML_Print_Metrix walks that tree recursively, but
Print_Metrix still uses the linear Node_Metrix. We should change
Node_Metrix to work like XML_Print_Metrix for uniformity.

Instead of passing in a per-file action, pass in a class-wide object with
Init, Final, and per-file action primitive operations. This allows to
store tool-specific information (in this case, the global XML output
file).