Implement lines metrics.
Implement the various lines metrics. Add new
METRICS.Line_Counting package to precompute a data
structure that can be used to efficiently compute
the number of lines with a given property in a
given range of lines. Add new Lines_Code_In_Bodies
and Num_Bodies metrics, used for computing
Lines_Average. Mimic various quirks of gnatmetric
output (e.g. the Average_Complexity_Metrics is not
indented like the others).

Implement printing of XSD (the schema). Put the
name of the schema in the XML file.

Clean up the Context