lalpp: Clean up predefined names
Move most calls to Intern that use a static string into
Predefined_Symbols, so they can be shared, and so they
can be initialized during library-level elaboration.
This was not possible before, because of issues with
initialization of Namet.

This allows to move Is_Op_Sym_With_Letters and
Op_Sym_Table to immediately within the body of
Pp.Formatting.

It also allows Gen_Plus and Gen_Minus to be of type Symbol.

Change "Intern (To_UTF8 (...))" to "W_Intern (...)".

P610-018