lalpp: Interfacing with GPS
Add Preserve_Line_Breaks (--preserve-line-breaks) command-line switch,
to be used by GPS to reformat text without changing line breaks.
This is not yet implemented; just the interface is there.

Add In_Range and Out_Range parameters to Format_Vector, for use by
GPS when the user wants to format just part of a file.
This is not yet implemented; just the interface is there.

Remove File_Name parameter, which was used for printing error messages.
Instead, pass the messages back to the client in the Messages vector.
This part is fully implemented.
Q824-031