lalpp: brackets encoding vs. string literals
This patch fixes a bug in which if we have a percent sign
inside a string literal, followed by a comment later in
the file, and [" within the comment, lalpp would get
confused and interpret the [" as the start of a
brackets-encoded wide character. That is wrong,
because brackets encoding is not supposed to be
recognized inside comments.

Fixes QB24-006