lal-tools: version numbers
We no longer build lalpp and then copy it to gnatpp;
we build gnatpp directly. Same for gnatmetric and
gnatstub.

Fix --version and -v switches to print what the ASIS-based
version prints.

Change tests to avoid -v switch (which now includes a
copyright year, so it will change).

As discussed in QC01-028.
Fixes P912-011.