lalpp: --indent-named-statements switch
The new switch causes named block and loop
statements to be indented relative to the
name.

Make named block and loop statements uniform
without the new switch. Old gnatpp was putting a
newline after the colon for loops, but not for
blocks.

Correct case of labels when (e.g.)
--name_mixed_case is given.

Fixes N312-029