lalpp: implement the --dictionary switch
The missing piece was that Scan_Dictionary was
not being called. We now have Scan_Dictionaries,
and call it.

The ASIS version had a kludge where Scan_Dictionaries
was being called late, and multiple times, because
Namet got destroyed when reading ASIS trees.
We no longer use the Namet package, so we don't
need that anymore.

No new test is needed, because we already have
tests for this feature. They are currently not
actually testing anything, because we're doing
case insensitive diffs. I tested it by hand,
and the other tests will be testing once QC28-009
and case sensitive diffs are enabled.

Fixes QC28-006