lalpp: greedy mode
Use the "greedy" algorithm, which packs as many
subtrees on the same line as possible. This seems
to look much better. The old algorithm was: if we
enable a soft line break at a certain level, we
enable all soft line breaks at the same level
(within a line we are splitting).

Fixes G905-011
Fixes N428-004