gnatmetric: Add <config> to the XML output,
...which maps XML names to text names.
Add a switch --no-xml-config to turn that off,
so we don't have to update a lot of test.out
files (just update the test-running script).
A test of this feature is in Q320-051.metric.c2-9.

Remove calls to To_W_Str, because for very
large files, it causes stack overflow.

R405-032

Change-Id: I4581d29d95546e5ed5374f8e9f014c8d3f436b6f