lalpp: improve formatting of formal parts
Formal parts were being improperly formatted.
This is now fixed. The main change was to change
the template in Do_Params to "?[# (~;#1 ~)]~"
so that the soft line break after each ";" is
considered to be more nested than the one
before "(". This ensures that the ones after ";"
will not be enabled unless the one before "("
is enabled.
Fixes R302-017.

Change-Id: Iba0fea73c142c8c24a67e3c3a7174a790cb0e6a0