lalpp: convert to using tokens instead of characters
More work on this. All phases are converted to new
scheme. Future plan: remove the old code.

Remove Token_Separator and related support for no-insert-line-breaks.
That feature will be much easier once we are fully converted
to token-based processing.

Use a named access type for the discriminant of Tokn_Cursor,
because an access discriminant doesn't allow default-initialized
objects, and we need to store these in Line_Breaks. Defaulted
access discriminants aren't allowed for nonlimited types.
This requires changing a lot of code to pass Lines_Data_Ptr
instead of Lines_Data_Rec.

Instead of appending new tokens to a temp token vector, and
then moving that to New_Tokns, we move New_Tokns to a temp
vector first, and then append tokens to New_Tokens. This
is necessary to get correct Tokn_Cursors stored in the
Line_Breaks table in Collect_Line_Breaks. We do it the
same way everywhere else for uniformity, even though
it's not strictly necessary.

R613-021

Change-Id: I1503518ce9bccd1331e7da7ee66c905f84b01d6f