gnatstub: correct output files
Use correct file names and directories
for subunits.
This should fix R809-027, except if a
Separate_Suffix clause is given in package
Naming in the project file.

Once R809-026 is fixed, then Separate_Suffix should work,
but we still need to test that.

Add target to Makefile to run utils-var_length_ints-test.

Change-Id: Iae73544f684c3d47f8d8d524ca2caecd06174d02