gnatstub: pragma Import and incomplete types
(Needs_Completion): For package specs, walk the
visible and private parts to determine if anything
in the package spec needs completion.
Return False for imported subprograms.
Return False fo incomplete types, which are
handled separately.

(Generate_Subunit_Start): Do not generate
"separate (parent)" if we are nested.

(Walk): Handle incomplete types separately, so
the completions appear before procedure bodies.
Do not generate a completion if there is one
in the spec; only stt-incomplete types need a
completion.
For package body stubs and protected body stubs,
walk the declarations in the corresponding spec.

(Per_File_Action): If the user requests a stub
for something that does not require a body,
give an error.

R807-021
R808-032

Change-Id: I6dd3c707ab806d20a2e0b2530fecc540f8cc16fd