gnatstub: --update-body switch
GPS can use the --update-body=N switch to tell
gnatstub to create a body for the subprogram on
line N of a spec.

Remove --comments-unchanged switch from the
invocations of gnatpp, because we want
generated comments to be formatted.
This affects laltools-testsuite/tests/stub/GC19-029,
but that was broken already. We should update
it to use "--!pp off" to turn off pretty-printing
in the header. This allows all gnatstub test
to pass with the --strict-whitespace-diff
switch passed to testsuite.py.

Remove the --syntax-only switch, because name
resolution is working well now.

Add the --comments-fill switch. This has no
effect on any tests, but seems like a good
idea because that's the recommended switch.

Use 'Access instead of 'Unchecked_Access;
that's better style. That code used to be
nested, so 'Unchecked_Access used to be
necessary.

Add --initial-indentation switch to gnatpp.
It does not yet fully work (it doesn't properly
indent the second and subsequent lines of a
block comment, and it puts a spurious line
with trailing spaces at the end).

Part of QC22-027

Change-Id: I2a7a9ecc5c40f35d0d3e327281cffa952995ce59