gnatstub: --update-body switch
Improve error handling. Improve formatting of the
output Ada code.

In Rec, when "Search_Line is not in Sloc",
always Update_Body_Line if there is a body;
otherwise we skip things like procedures
with bodies, and the output is in the wrong
order.

Disable the backup file.

Add more test cases in
testsuite/tests/stub/QC22-027.

...and make sure the test.out reflects that $LC_ALL=C .
Otherwise, wildcards expand in a wrong order.

QC22-027

Change-Id: Ic86677c5959f222e875196a2edaf65472e3393ba