gnatmetric: coupling metrics
Add the necessary name resolution to complete the
implementation of the coupling metrics. This exposed
some incorrect output from the ASIS-based version.

Fix coupling metrics to handle "limited with"
more uniformly. In particular, make them transitive
for all coupling metrics, not just some. Remove
Limited_Depends_On, and just use Depends_On.

Update hierarchy coupling to avoid counting
dependences within the hierarchy.

Add assertions.  For example, fan-in and fan-out
are symmetric; the total of all X_Coupling_In
should equal the total of all X_Coupling_Out (for
the 4 different X).  See for example GB13-005,
where this is not true in the ASIS-based version.

The ASIS-based version was using a strange
algorithm to determine whether to print 0-valued
coupling metrics. For example, KB07-010 was printing
<metric name="control_fan-out_coupling">0</metric>
but nothing for fan-in.

Fixes RA01-047

Change-Id: I2b0e3ba3f60752c5a1f4acdb8d19f8491c922c4f