gnatmetric: all_types, param_number
For All_Types, do not count types with
completions twice.  Do not count generic
formal types.
Move printing of Computed_All_Types to the
correct place to match the ASIS-based version.
Do not print complexity metrix if the
--no-local-metrics switch was specified.
Fixes RA01-053
Fixes MB05-004

We used to compute Param_Number and friends by
taking action on Ada_Param_Spec nodes -- increment
the counts in the innermost enclosing Metrix,
which should be the corresponding subprogram.
But that doesn't work for subprogram instantiations,
where we have to reach over to the generic subprogram.
It's also kind of kludgy, because some parameters
are not counted (parameters of formal subprograms,
access-to-subprogram types, for example).
So we now attach the action to the subprogram
node itself, and walk down to the parameters
to count them (total and by mode).
Fixes RA07-006.

Change-Id: Ia6c5737fedc20278ced12f8e36d29f4bc0736ee0