gnatstub: Reparse if --update-body
Revision e5b94a476c90c9968c192284b49d13ac5273bb5e
changed the call to Get_From_Buffer to Get_From_File,
to work around a bug, and to improve efficiency.
This patch refines that, so if gnatstub is
called with --update-body, we pass Reparse to
Get_From_File.

Part of RA04-024

Change-Id: Iae8052bd5a154e5a5d1144ab2d543be0ab0560af