gnatstub: Allow directory name
The ASIS-based version of gnatstub allows
a directory name on the command line
to specify the output directory.
That is, "gnatstub x.ads dir" means
the same as "gnatstub x.ads --output-dir=dir".

This feature is not documented, but
unfortunately, GPS uses it.  So this
patch implements it.

Fixes RA17-035

Change-Id: I59cf529e77f4a678fd18dd7c07925d9f317bd9ba