gnatstub: Fix QC22-027 regression on windows
The introduction of Copy_Command_Line introduced
a dangling pointer in Set_Arg in Other_Switches,
because it doesn't do a deep copy. Rather than
introduce a complicated and inefficient deep
copy, we simply avoid freeing the string in
Set_Arg. It is called very rarely, so there
is no problem with a small memory leak.

Avoid uninit var in Pp.Buffers, in case the input
string is empty.

Misc cleanup.

S109-028

Change-Id: I82ddcb825477a7e740b3a58050f0f8a499015695