gnatpp: Add support for preprocessing
Add support for preprocessing.  A line starting
with "#" is now considered a token --
a Preprocessor_Directive.  These are handled
similarly to Whole_Line_Comments, and are
inserted into the token stream (New_Tokns)
by Insert_Comments_And_Blank_Lines.

This works only if the input is syntactically
legal after deleting all the preprocessor
directives.  This is a deliberate limitation,
as discussed.

R504-013
RA19-026
S213-005

Change-Id: Id4f6a00180fcfa88adaf261fe2896805b93915e6