gnatstub: Generate body for --subunits
Change the --subunits switch so that it works on a
package spec. Instead of giving an error (as the
previous version did), if the package body is
missing, it is generated, containing Ada stubs
for all the subprogram specs in the package spec,
and then the subunits are generated as if
gnatstub had been called on this newly-generated
package body.

Fixes QB30-025

Change-Id: I7980ed133c7a6fa80e08d6f76d7ce8bfc43b9764