gnatmetric, gnatstub: Enable aggregate projects
Support for aggregate projects was implemented for gnatpp
under P510-022. That was copied from ASIS support for
aggregate projects in gnatcheck, which was enabled only
for gnatcheck, so the support in libadalang was enabled
only for gnatpp. This patch enables the same support for
gnatmetric and gnatstub.

Fixes RC12-009
Fixes S410-054

Change-Id: I543953f2c1c163d1e09a87ed30c21b83bebc1148