Separate regular and WIP programs in build system and testsuite
This commit introduces the concept of WIP program: work-in-progress
(WIP) programs are being developped, so depending on the context we do
not want to build/test them. For now, only gnattest is a WIP.

In order to do this, this change introduces switches in the build system
(the LALTOOLS_SET make variable and GPR external) and a new testsuite
command-line option (--no-wip). By default, WIP programs are built and
tested. In order not to build them, pass LALTOOLS_SET=no-wip to make(1),
and pass --no-wip to testsuite.py not to test them.

Change-Id: If7242a887d43ab202afc6972218679c5e7be2c2b
Depends-On: Id2a2afba4506808858a0bb79154bb6b76a2ffcdd
TN: S729-028