gnatpp: Improve --vertical-named-aggregates switch
Adjust the cases where aggregates can be formatted vertically. Allow
vertical aggregates as the default value in a component declaration,
analogous to the initial value in an object declaration. Do not format a
singleton aggregate vertically, except if the single component is itself
an aggregate.

Improve documentation and clean up code.

Fix Before_Indentation so it doesn't index Syntax_LBI
out of bounds.

T604-015

Change-Id: I9feaa0d4ce6d90dabde44df7174b3bd69bbbb843