SA22-035 Adjust testing infrastructure for import refactors
As part of Makefile, build the project containing the test driver.

Rename the project to something slightly shorter. Use a 'bin' directory
that can be shared by several testsuite drivers in the future.

As part of the testsuite driver, place this 'bin' directory in front of
the PATH before executing the testsuite, so that the (relocated) test
scripts can find the testsuite drivers.

Adjust the test.sh accordingly.

Change-Id: I2dee63fe1fc40a12b5732e7b76dddaa2acb1ec48