SA22-035 Generics and renames for refactor imports
Declarations from instantiations of generic package are found and
with clause / prefix suggestion is given.
Packages that rename a package that is part of a suggestion are now also
included in the suggestions list.
Suggestion list is now sorted first by the with clause and then by the
prefix. Test outputs were updated accordingly.

Change-Id: I80b26c964b1f61802297665954978b38439ab05e