Improve formatting with --vertical-case-alternatives
If --vertical-case-alternatives is given, insert a hard line break after
the "is" of a case expression, and between alternatives. In addition, if
the context has a soft line break, treat it as hard, and skip a
following space, if any. Use the same treatment for
--vertical-named-aggregates, which allows us to remove some of the
"Vertical_Agg_Alt" templates and corresponding special cases.

Misc cleanup.

T730-039

Change-Id: I9a3c13e8a6bc4b1ecddda84dcbe99a18343bf301