gnatpp: Use compact format for binary operators
Ignore the --no-compact switch given with respect to
binary operators. That is, format things like "A & B & C"
as if the default --compact mode is used, whether or not
--compact or --no-compact is in place. In particular, we
put as many operands as will fit on a line.

Bin_Op_Count is a count of the nesting level of binary
operators. Increment and decrement this in Do_Bin_Op.
Record the current value of Bin_Op_Count in each line
break that is created during the first pass. (We always
use zero during later passes.) In Split_Lines, when we are
looping through all the soft line breaks at a given
nesting level, if Bin_Op_Count iz nonzero, do not enable
the line break unless it is necessary to keep the line
short enough.

Fixes T731-010
Fixes T718-004

Change-Id: If0354da748bb79bdd0e3b68587e4ab3fb559ee1c