Gnatcov integration: Overhaul Makefile generation
As part of the integration of gnatcoverage in gnattest,
the Makefile generated by gnattest only had the coverage rules
in certain situations, and was missing a lot of use cases.

It is now generated in all cases, and automates coverage analysis
of the project under test in all situations.

TN: U115-024
Change-Id: I813ecb427c039dcbbdd677bc6c27457750ebd072