U511-009: laltools refactor import improvements
Add Get_Insert_With_Location function: it allows to parse
the list of with statement in a file and find the location
where to insert a new one if needed (and not crash on use_clause)

Fix the handling of nested package inside a main unit.
Fix incorrect codefix when no with clause: trying to prefix the
entity by itself.

Change-Id: Idceaec5122c623ef1891676a3b31dd8c81becad4