Reconnect --strict switch in gnattest
Reconnect --strict switch.
Add Per_Invalid_File_Action to the driver which is called
for units with diagnostics, this allows to properly detect
cases of invalid units and set gnattest exit status accordingly
when --strict is specified.
Also check for GNATTEST_STRICT env variable, and when it is set
to "TRUE" count it as --strict specified. This will allow to
launch gnattest in strict mode in all testsuites without manually
setting the switch in each test and catch potential regressions
once protection against possible lal crashes on incomplete and or
incorrect sources is added.

Depends-On:  I04a93bca108975732c5f9f95edb86448bc19235c
Change-Id: I65fab96969eff7091c2e78d4618820acd351602e
TN: U808-002, RC01-008