U813-002 Updated default casing of number decls and const decls
Default of number declarations: if not defined, use casing of constant
declarations.
Defualt of constant declarations: if not defined, using casing of non
constant declarations.

Before this patch, the default option was "as declared" however, it was
using the casing of non constant declarations. This patch also fixes
this problem.

Updated previous testcase and added a new one for number declarations.

Change-Id: Ic8710a9cd994be28e36e2d1345e51a6561575515