Process subprograms from private part
Fix gnattest to create skeletons for subprograms declared in the private
part of UUT specs. Do not replace controlling type name with corresponding
root type name when controlling type is declared in the private part
of the package. Such situation is treated as non-dispatching by gnattest
and name clashes of test routines were possible.

Change-Id: Idb70d64545973d3fd1b3e643ea942272f1b33780
TN: UA15-049