Make the return type of the translation a refcounted type
This will make memory managment much easier for the user.
This change also introduces memoization for the Translate
functions. With this, a specific type declaration is only translated
once, and every time it appears (be it as a formal param of a
subprogram or nested in a composite type), then a reference to the
same translation object will be returned.

no-tn-check

Change-Id: I92bd906294156489afcc0aae2f0d1e3dede17e74