Add partial record type representation
This change adds translation for record types, although it is not
final. Missing are discriminated record types that are constrained,
as well as discriminated record derived types that "rebind" some of
the discriminants of the ancestor part (See RM 3.7 (18)).

This change also instroduces reference counting for translation
results to simplify memory managment as well as a memoization of
the translation results to ensure a same type referenced in mutiple
parts of the same code only gets translated once.

no-tn-check

Change-Id: I10601cfd1b33fb238fc9d8dd400a265a195c4967