TGen: Add support for decimal fixed point generation
Generation for those types is done by dividing the range of the
fixed point type (if any) by the size of the delta, to obtain the
number of elements actually present in the type. An Integer value
is drawn from that range and the generated value is set to the
delta multiplied by that integer value.

TN: UA28-013
Change-Id: I1efc189ab4970f71cf5bb9182cc97a79a4c3ae08