TGen: Support mutable discriminants
When it is at the top-level, a type with mutable discriminants is
handled as other record types with discriminants, ie. the discriminants
are stored in headers which should not be mutated with the rest. If it
is used as a component type, the discriminants are marshalled with the
rest and some padding is added so that even if the discriminant is
mutated, we never read into the next component.

To do that, we generate Size and Max_Size functions for all types. The
Size function computes the number of bits used to marshall a given
value. The Max_Size function computes the maximum number of bits that
can be used to marshall any value of the type. The Max_Size function
takes as parameters the bounds for arrays, and bounds on the
discriminant values for records. The result of the Max_Size function
for type with mutable discriminants is computed once and for all during
the elaboration of the package. The Size function is computed again
for each marshalled value. We could consider merging it with the read
or write of the object if we think it would be better for efficiency.

Change-Id: I7ea2213933e016614e0afaef0859202a0ff182ab
TN: UB03-008