Gnattest: Fix handling of declarations with multiple defining names
A recent update in Libadalang (intentionally) created new exceptions
when querying the "ghost" status for declarations that define multiple
entities, such as "A, B : Integer", as one entity may be ghost but not
the other.

Query the ghost status directly on each defining identifier node instead
of on the declaration node to avoid this problem.

TN: V520-030
Change-Id: I734f57b462dbe99e5c0812d45f53fd3fd518ab1e