Add support for Big_Reals in TGen
Now that LAL uses Big_Reals during static evaluation of float
expressions, use that as well in order not to loose any precision.

This is a reapplication of Iada75de325093716c8df534d5c17a77580b5d853
which used a function defined in Ada.Numerics.Big_Numbers.Big_Reals
not yet present in the fsf gnat runtime, and thus breaks the community
build of the als. This function is now implemented in TGen, we'll be
able to remove it when the From_Universal_Image function makes it to
fsf gnat.

TN: UA28-013
Change-Id: If6d78097b7c7d80ba478917b8d44e698940a51ba