V801-024 Add logger and exception handlers
Add a default public Trace_Handler that can be replaced by a user of the
refactoring library.
Add exceptions handlers to all Is_<Refactoring_Tool>_Available and log
any exceptions. The rationale to not propagate the exception here is
that these functions are expected to be called on invalid code, therefore,
exceptions can be raised. In that case, False should be returned and the
exception should be logged. Exceptions raised on valid code are
most likely bugs in these functions. These can only be identified by
looking at the logs.
Add exceptions handlers to all <Refactoring_Tool>.Refactor functions and
log any exceptions. In this case, No_Refactoring_Edits is
returned. These functions are expected to be called on valid code, so
any exceptions are most likely bugs.

Change-Id: I84733d8607619e65e6cd6425dc8179b2be8f9191