TGen: split parts needed for value generation from the rest
For dynamic generation needs, we will need to generate code that will use
parts of TGen, notably the type representation and the strategy
generation. As we don't want to have any dependency over libadalang in
the generated code, split the core of tgen value generation (consisting
of the type representation + the strategy generation) from the type
translation.

As this is quite the overhaul, also remove bits of code that referred to
the old dynamic generation scheme, which we won't use anymore in the
future:
   * Remove the old templates
   * Remove function generators
   * Remove dynamic strategies

Note that the refactoring is not exhaustive. At some point, we will have
to get rid of the "Is_Static" field present in every type representation,
as we will end up relying purely on the dynamic generation.

TN: UA28-013
Change-Id: Ibd862f93a8a3e592591671aaf9a3d52dacc3d61a