TGen: preliminary work for dynamic generation
This commit introduces initial (partial) support for dynamic
generation. We now use a JSON value kind in the random generation
engine to have a more structured representation of values, and to ease
up the generation of marshallers / unmarshallers converting from the
random generation data representation to actual Ada types.

Part of the work still needs to be done, i.e. using these generators in
the generated gnattest harness.