TGen: Fix code generation bug in marshalling lib
When generating the support library for some type T, TGen gathers the
types on which T directly depends on and "withs" the support package
of those types in the package being generated. TGen was not including
the support package for anonymous types (in record components for
instance), whereas they are required. This change fixes this.

The change also moves the "with" of depended-on support units to the
body of the currently generated support package to avoid circular
dependencies.

TN: UB03-008
Change-Id: Id9628e95c7a41409e3dbd73c05c98a98ff2cdcc1