Fix wrong testsuite discriminant
Commit 26ceb1d0d9b50f3afb1a365cce874ea675d8ca77 introduces a '32bits'
discriminant allowing to XFAIL tests on 32bit hosts, but using the
string '32bits' was not possible as a python identifier cannot start
by a digit character and thus the evaluation of the XFAIL condition
failed.

This change adresses this by making the discriminant name start
with a letter.

TN: VC07-013
Change-Id: I3815e396adaccb19965d3544b8f432b87316db6d