Add multiple functionalities to gnatrefactor array_aggregates
Refactor the algorithm used to find aggregates. Only try to resolve the
expression type of an aggregate (helps with memory consumption).
Add --sources,-S switch to allow the user to select the sources he
wishes to refactor. If combined with --project,-P then the tool is able to
resolve the type of each aggregate and not miss any. Otherwise some array
aggregates might be missed.
Add -X switch to allow the user to defined scenario variables.
Add --from-gnat-warnings switch to process array aggregate obsolescent
syntax warnings produced by GNAT. This is has the advantage that it uses
less memory, since that for each warning a new LAL context is created. This
is possibly slower, but at least does not suffer from huge memory
consumption when many sources are analyzed.