Testsuite: add the new "-k" switch where needed after a LAL change
A recent commit in Libadalang changed the default behavior for command
line tools based on the Libadalang.Helpers.App package: when a source
file is missing for name resolution, while previously the app just
printed a warning, it nows terminates with an error. Add the new "-k"
command line switch to restore the previous behavior in tests where this
matters.

To be submitted with eng/libadalang/libadalang!1233
TN: VB23-023