Fix precision of real types in the JSON encoding
If V is a value of a float type F, then F'Value (V'Image) is not
necessarily the identity function because of the loss of precision in
the Image function. As we don't want to lose precison in the JSON
marshallers, we will represent the value as an exact fraction and decode
it from a string fraction.