Do a case insensitive comparison when looking for collisions
Introduce a new New_Name_Type (and an access to it) to hold the original
and new name, and minimize string copies.
Introduce Specific_Problem_Finder_Result to express that a specific
problem finder algorithm can return a problem or not. This also removes
the Dummy_Problem that used to represent no problems found.
Add multiple Pre contract to type constructors to better express
contraints.