Test: Always load JSON test if present
This change makes it so that JSON tests will always be loaded in
the harness if they are present in the predefined test directory
(obj/gnattest/tests/JSON_Tests).

The predefined JSON_Test directory location can now be changed with
the currently undocumented --serialized-test-dir command line switch.

Part of V315-014