TGen: Introduce test-case serializers
The test-case serializers are support procedures, generated
by libgen, including the same parameters as the subprogram for
which it is being generated. It creates or augments a JSON value
representing test cases for a whole unit, adding a new test-case
under the correct subprogram entry. This JSON value can be directly
serialized to a file and be used by gnattest to generate a unit test
harness from it.

Still missing are the expected values for the out parameters and
return value.

A sample use of the new functionality can be found under
testsuite/tests/test/testcase_json_dump.

TN: V315-014