Gnattest: Allow JSON tests to be loaded without test gen
The harness generated from serialized JSON tests needs to have
a tgen_support library to compile. Prior to this commit, the
tgen_support library would only be generated when generating tests.

With this change, the libgen context is always created, and we also
emit the tgen_support library as soon as we detect that JSON tests
are loaded.