TGen: Expose new API defining the filenames for serialized tests
This change adds two new functions in TGen.LAL_Utils which, given
a subprogram declaration node, will return the simple name of the
file in which the serialized tests should be stored / retrieved from.

For binary tests, the current convention is to have a single test per
file, so the filename includes the name of the subprogram under test,
as well as a unique identifier to handle overloads.

No behavior change expected.