TGen: Generated helpers for private types as private subprograms
Types only declared in the private part of a package can only be
referenced in that private part, or in the private part of its child
packages. The helper subprograms generated in the support library
were all generated in the public part, which violated visibility rules.

This change moves to the body (which is part of the private part) as many
implementation details as possible for the value library, and ensures
marshallers are generated in the private part if needed.