TGen: Fix bug in type visibility computation
In order to determine whether a type had any public view, TGen relied
on the chain of semantic parents, as returned by P_Semantic_Parent.

This does not work in our case as for a given public child package,
the property returns the private part of its parent package.

This change instead stops traversing the semantic parent chain as soon
as a library level package declaration is encountered.