TGen: fix generation of marshallers for scalar type with private ext
TGen used to generate invalid code (not respecting visibility rules) for
the code of marshallers of scalar types with a private extension. The
visibility violation occurred when instantiating the `Size` (non
exhaustive) subprogram, which needs to compute the <scalar_type>'First
as a default parameter value.

To fix this, we actually generate a subprogram to get the 'First (and
'Last) of a scalar type, that is defined as a private extension when the
type is private.

Also do some refactorings to simplify parts of the code.