[TGen] remove imprecise preconditions
We check against the root type declaration before invoking the right
translate subprogram. In this translate subprogram precondition, we
check the type of the base type declaration, which is not right as the
root type can be private, meaning that the base type will not be e.g. a
scalar view if the root type is.