[TGen] Do not support subprogram with a global input unsupported
Before this patch, TGen used to skip global inputs that were not
supported - with a verbose warning, but this is not transparent to the
user. We should rather consider the subprogram as unsupported if one of
the global types is not supported, just like we do for parameter types.

Adapt the test baseline accordingly.