Test: Introduce testsuite minimization
This first iteration of testsuite minimization uses gnatcov to
remove any generated (JSON) tests that do not contribute anything
to coverage on top of the previously executed tests.

This first iteration may not work well in all cases for MCDC coverage,
as it relies on the reports from gnatcov. The later does not necessarily
report a coverage  increase when a new evaluation of interest happens for
a given condition, thus some tests may be discarded as they individually
do not increase the coverage metrics, but taken together more conditions
might have been covered.