Test: Fix generation of mapping file after generation
The minimization file launches an instance of gnattest as a sub-command
to re-generate a harness once all the redundant tests have been removed.
This also generates a up-to-date mapping files, used by gnatstudio to
find the test-cases.

Ensure that it is not overwritten by the top-level gnattest call,
which still contains mapping information for tests that may have been
deleted.