Test: Use the coverage level defined in the user project for minimization
If the --cov-level switch is passed, inspect the user's root project
Coverage package to determine the coverage level to be used during
testsuite minimization.

In other gnattest uses, this is taken care of by the Origin_Project
attribute, it simply can't be used together with the manual dump
trigger of gnatcov, essential for the minimization feature.