Testsuite: Introduce shared resource mechanism
It isn't common for multiple tests to need shared resources.
This patch introduces a shared resource mechanism. It allows a test
to declare a set of resources that will be copied in the test directory.
These resources are put in a subdirectory of the `testsuite/shared` directory
for those that can be publicly exposed, or in the `shared` directory of the
laltools-testsuite repository for the resources that must remain internal
to Adacore.

A test.yaml script can then declare the resources it needs by adding a
`resources` entry in the yaml file:

```
resources:
    resource_subdir_name: destination_subdir_1
    internal/resource_subdir: destination_subdir_2

```