TGen: Fix wrong code generation for wrappers in case of name conflict
When a subprogram in a given package has a parameter of the same name,
the later shadows the former, making it impossible to reference the
original subprogram by its fully qualified name in the wrapper.

This is fixed by introducing a renaming for the package in which
the subprogram is declared.