Test: add optional -o command line switch for test input dumping
This allows other tools using the test runners to not pollute the user's
project space with artifacts, by specifying the directory in which the
test input files should be created.

The only recognized switch is -o (no --output); if not specified the
test inputs are dumped in the directory of execution of the test
harness.