Fix API for diagnostics / unsupported types
The API is friendlier if the diagnostics are returned as a list of
strings, rather than a big concatenated string, as it allows the
endpoint user to configure the displaying the way he wants.

This commit does not change the behavior of the tool, so it does not
require additional testing.