TGen: Fix visibility issue for generated support library and test harness
When a subprogram or a type has the same name as the package in which is
is declared, it is not possible to reference it by its fully qualified name,
as the package name is shadowed by the local entity.

Fix this by introducing a rename of the standard package (in case it is shadowed
itself), and prefixing all entities by that renaming. This introduces some
complexities around the management of the fully qualified names, as LAL is
inconsistent in whether the "standard" package is included in the qualified name.
This is only the case for predefined entities, so some additional special processing
is introduced to account for this.