Test: fix processing of aggregate projects
The laltools framework supports aggregate projects in the sense
that they will be processed sequentially, invoking the tool in a
subprocess on each of the aggregated projects. It also however calls
the Init and Final tool methods on the aggregate project itself, which
is not necessary in gnattest's case, and even used to lead to a crash due
to call to GNATCOLL.Project primitives that were illegal on aggregate projects.

This fixes the crash by skipping the Init and Final call of the test tool when
the aggregate project is being processed.