Test: fix handling of extended projects in stubbing mode
When generating stubs, gnattest only processes the most extending
project when extending projects are present. It also however keeps
track of the project inclusions, which did not account for project
extensions and thus resulted in references to projects not present
in the internal project table.

This is fixed by always using the extending most view of a project
when computing dependencies.