Add special testsuite driver to execute gnattest+TGen on gnatfuzz tests
This driver invokes gnattest in test generation mode and instrumemts
the sources to dump the test cases and records the test harness execution
as well as the number of test cases dumped in the baseline. The purpose
of this is mainly to benefit from the codebases in the gnatfuzz testsuite,
and to anticipate any new issues the gnatfuzz team may encounter after a
laltools change.

A new CI job is also introduced to run these tests specifically when
TGen files are modified.