TGen: Change parameter name of generate procedure to avoid conflicts
The procedure to dump tests inputs to a JSON file has the same parameters
as the subprogram from which it was generated (with a TGen_Marshalling_ prefix),
and an additional TGen_Marshalling_Origin parameter, which leads to conflicts
when test tested subprogram has a parameter named Origin.

Change the prefix of the predefined parameters (Origin and Unit_JSON) to
ensure no conflicts can happen.