[TGen] support top level generic instantiations when dumping test inputs
The issue was two-fold: when instrumenting to dump test inputs, gnattest
used to inspect the generic package in addition to the generic package
instantiations, which resulted in generating non-generic child packages
of a generic package (which is not allowed by the standard).

Also, when dumping test inputs without generating test case vectors,
gnattest did not pass the right value for the
Is_Top_Level_Generic_Instantiation formal of the Include_Subp
subprogram. This formal can actually be computed in the body of
Include_Subp, so do that rather than asking the caller to provide this
information.