Test.Instrument: Fix instrumentation of operators
The instrumentation process in gnattest introduces a wrapper around
the original user code. This wrapper mean we generate a new subprogram
name from the original subprogram, by adding a _GNATTEST at the end.

When the user suprogrma is an operator, this used to result in illegal
code.

This change fixes this by intorducing a mapping for all Ada operators
when generating wrapper subrpogram names.